ASCollection: implement IEnumerable (#110)
Resolves #43

It does not implement the generic IEnumerable<T> as OrderedCollection is no longer generic. If we still desire it to implement a generic IEnumerable then we'll have to revert or partially revert this, but I think implementing as IEnumerable<Linkable<ASObject>> should hopefully not be a problem since that's what the Entity's LinkableList is essentially doing.