Allow entities to specify a custom context
Entities and Models can now specify an alternate or extended JSON-LD context. In normal usage, this is done by overriding `IASModel<TModel>.DefiningContext` to provide a static, immutable instance of the context definition. This is mirrored via a pivot in `ASEntity<TModel, TEntity>.DefiningContext`, which allows access from either side.

When an entity is added to a TypeMap, its context is merged into the mutable instance at `TypeMap._ldContext`, and exposed (immutably) through `TypeMap.LDContext`. Duplicates are ignored, ensuring that this property always contains a minimal superset of all contexts in use.

Implements #95